Add Tree & AncientBuidling PCG Gen Tools

+++ Source/Terrain/terrain_3d_material.cpp
@@ -926,7 +926,8 @@ void Terrain3DMaterial::set_projection_enabled(const bool p_enabled) {
 }
 
 void Terrain3DMaterial::set_max_regions(const RegionMaximum p_max) {
-	SET_IF_DIFF(_max_regions, RegionMaximum(CLAMP(closest_power_of_2(p_max), 64, 1024)));
+	// Explicit width: godot-cpp master overloads closest_power_of_2 for 32 and 64 bits.
+	SET_IF_DIFF(_max_regions, RegionMaximum(CLAMP(closest_power_of_2(uint32_t(p_max)), 64, 1024)));
 	LOG(INFO, "Set max region count: ", _max_regions);
 	_update_shader();
 }
